jwt validation function improvements, verify admin functionality
