+++ server/controllers/usersController.js
@@ -15,6 +15,7 @@ const login = async (req, res) => {
     };
     const { password, _id, ...user } = DB_RESPONSE;
     const JWT_accessToken = createAccessJWT(user);
+    console.log(JWT_accessToken);
     res.header('Authorization', JWT_accessToken).send({ success: `[${user.email}] was logged in successfully`, userData: user });
   } catch(err){
     console.error(err);
@@ -25,18 +26,19 @@ const login = async (req, res) => {
 };
 
 const autoLogin = async (req, res) => {
-  // extract authetication header from request, contains the JWT token in format { Authorization: Bearer <token> }
-  const authHeader = req.headers.authorization;
-  // if thea authentication header is empty, send a message that user needs to re-authenticate
-  if(!authHeader){
-    return res.status(401).send({ error: 'Token does not exist. Please log in again.' });
-  };
-  // extract the access token from the header and validate
-  const accessToken = req.headers.authorization.split(' ')[1];
-  const verifyResults = validateJWT(accessToken);
-  if(verifyResults.error){
-    return res.status(verifyResults.status).send(verifyResults);
-  };
+  // // extract authetication header from request, contains the JWT token in format { Authorization: Bearer <token> }
+  // const authHeader = req.headers.authorization;
+  // // if thea authentication header is empty, send a message that user needs to re-authenticate
+  // if(!authHeader){
+  //   return res.status(401).send({ error: 'Token does not exist. Please log in again.' });
+  // };
+  // // extract the access token from the header and validate
+  // const accessToken = authHeader.split(' ')[1];
+  // const verifyResults = validateJWT(accessToken);
+  // if(verifyResults.error){
+  //   return res.status(verifyResults.status).send(verifyResults);
+  // };
+  res.send({ success: 'User was authenticated.', userData: req.user });
 };
 
 const refreshToken = async (req, res) => {
